+++ src/getStepStartStates.test.ts
@@ -74,7 +74,10 @@ import {
   STEP_RM_CDN_PROFILE,
   STEP_RM_CDN_ENDPOINTS,
 } from './steps/resource-manager/cdn';
-import { STEP_RM_BATCH_ACCOUNT } from './steps/resource-manager/batch';
+import {
+  STEP_RM_BATCH_ACCOUNT,
+  STEP_RM_BATCH_POOL,
+} from './steps/resource-manager/batch';
 
 describe('getStepStartStates', () => {
   test('all steps represented', () => {
@@ -138,6 +141,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_CDN_PROFILE]: { disabled: true },
       [STEP_RM_CDN_ENDPOINTS]: { disabled: true },
       [STEP_RM_BATCH_ACCOUNT]: { disabled: true },
+      [STEP_RM_BATCH_POOL]: { disabled: true },
     });
   });
 
@@ -197,6 +201,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_CDN_PROFILE]: { disabled: true },
       [STEP_RM_CDN_ENDPOINTS]: { disabled: true },
       [STEP_RM_BATCH_ACCOUNT]: { disabled: true },
+      [STEP_RM_BATCH_POOL]: { disabled: true },
     });
   });
 
@@ -256,6 +261,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_CDN_PROFILE]: { disabled: false },
       [STEP_RM_CDN_ENDPOINTS]: { disabled: false },
       [STEP_RM_BATCH_ACCOUNT]: { disabled: false },
+      [STEP_RM_BATCH_POOL]: { disabled: false },
     });
   });
 });