+++ src/getStepStartStates.ts
@@ -73,7 +73,10 @@ import {
   STEP_RM_CDN_PROFILE,
   STEP_RM_CDN_ENDPOINTS,
 } from './steps/resource-manager/cdn';
-import { STEP_RM_BATCH_ACCOUNT } from './steps/resource-manager/batch';
+import {
+  STEP_RM_BATCH_ACCOUNT,
+  STEP_RM_BATCH_POOL,
+} from './steps/resource-manager/batch';
 import { hasSubscriptionId } from '.';
 
 export default function getStepStartStates(
@@ -131,5 +134,6 @@ export default function getStepStartStates(
     [STEP_RM_CDN_PROFILE]: resourceManager,
     [STEP_RM_CDN_ENDPOINTS]: resourceManager,
     [STEP_RM_BATCH_ACCOUNT]: resourceManager,
+    [STEP_RM_BATCH_POOL]: resourceManager,
   };
 }